Update to PIRSF post-processing to enable multipart matches:
The local-local nature of HMMER3 means that some of the PIRSF entries
were under hitting compared to the HMMER2 version.  This is because
only one domain hit was being considered. This is a forward port of the
PIRSF changes to our code. There is now the need need to read in
and look for sets of results. These are then pass to a new
post-processing subroutine, which determines the total hit length.
A ratio of hmm/sequence coverage has since been added to ensure
that over 2/3 of the HMM is covered with respect to the sequence length.
(ish)